+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Fixed bug in x86 interval `OR` function ([#395](https://github.com/mkeeter/fidget/pull/395)),
   thanks [@Wulfsta](https://github.com/Wulfsta)
 - Bumped many dependencies
+- Add `Image::build` function to build an image from a `Vec<T>` and
+  `ImageSizeLike`, returning an error if the data size is incorrect.
 
 # 0.4.2
 - Change `depth` member in `GeometryPixel` from `u32` to `f32` ([#381](https://github.com/mkeeter/fidget/pull/381))